backend/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptstudio-python-backend"
version = "1.0.0"
description = "PromptStudio Python Backend with FastAPI, Mirascope, and Instructor"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PromptStudio Team"}
]
keywords = ["llm", "fastapi", "mirascope", "instructor", "prompts"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "mirascope>=1.5.0",
    "instructor>=1.3.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "python-socketio>=5.11.0",
    "websockets>=12.0",
    "redis>=5.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
